Added delete user from user_metada

+++ server/assets-v1/templates/cypress.config.js
@@ -22,14 +22,30 @@ module.exports = defineConfig({
             host: String(process.env.DB_HOST),
             database: String(process.env.DB_NAME),
             port: String(process.env.DB_PORT),
-            ssl: false,
+            ssl: process.env.SSL_MODE === 'true' ? { rejectUnauthorized: false } : false,
           });
 
           try {
             await client.connect();
-            await client.query(`DELETE FROM ${tableName} WHERE username = $1`, [username]);
-            await client.end();
-            return true;
+            if (tableName == 'users') {
+              // Retrieve user ID by username from tableName
+              const userResult = await client.query(`SELECT id FROM ${tableName} WHERE username = $1`, [username]);
+              if (userResult.rows.length === 0) {
+                throw new Error(`User with username ${username} not found in ${tableName}`);
+              }
+              const userId = userResult.rows[0].id;
+
+              // Delete data from user_metadata table using user ID
+              await client.query(`DELETE FROM user_metadata WHERE user_id = $1`, [userId]);
+              await client.query(`DELETE FROM ${tableName} WHERE username = $1`, [username]);
+
+              await client.end();
+              return true;
+            } else {
+              await client.query(`DELETE FROM ${tableName} WHERE username = $1`, [username]);
+              await client.end();
+              return true;
+            }
           } catch (err) {
             await client.end();
             throw new Error(`Database error: ${err.message}`);
